requirements.txt
Flask==3.0.0
Pillow==9.4.0
ezdxf==1.0.6
